+++ pkg/ports/ports_test.go
@@ -1,4 +1,4 @@
-package cmd
+package ports
 
 import (
 	"testing"
